requirements.txt
altair==6.0.0
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
branca==0.8.2
cachetools==7.0.5
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comtypes==1.4.13
et_xmlfile==2.0.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
folium==0.20.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==6.0.2
MarkupSafe==3.0.3
mysql-connector-python==9.6.0
narwhals==2.18.1
numpy==2.4.1
openpyxl==3.1.5
packaging==26.0
pandas==3.0.0
pillow==12.0.0
protobuf==7.34.1
pyarrow==23.0.1
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-pptx==1.0.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
soupsieve==2.8
SQLAlchemy==2.0.46
streamlit==1.56.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.5
xlsxwriter==3.2.9
xyzservices==2025.11.0
gunicorn==23.0.0
